requirements.txt
# AI and Image Processing
openai>=1.12.0
pillow>=10.2.0
anthropic>=0.18.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Social Media APIs
facebook-sdk>=3.1.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Optional: Video processing (for TikTok videos)
# moviepy>=1.0.3
# opencv-python>=4.9.0
